+++ cmd/kubectl-nifikop/nifidataflow/input/input.go
@@ -1,0 +1,36 @@
+package input
+
+import (
+	"github.com/konpyutaika/nifikop/cmd/kubectl-nifikop/nifidataflow/input/stop"
+
+	"github.com/spf13/cobra"
+	"k8s.io/cli-runtime/pkg/genericclioptions"
+)
+
+// options provides information required by clusteragent command
+type options struct {
+	genericclioptions.IOStreams
+	configFlags *genericclioptions.ConfigFlags
+}
+
+// newOptions provides an instance of options with default values
+func newOptions(streams genericclioptions.IOStreams) *options {
+	return &options{
+		configFlags: genericclioptions.NewConfigFlags(false),
+		IOStreams:   streams,
+	}
+}
+
+// New provides a cobra command wrapping options for "clusteragent" sub command
+func New(streams genericclioptions.IOStreams) *cobra.Command {
+	cmd := &cobra.Command{
+		Use: "input [subcommand] [flags]",
+	}
+
+	cmd.AddCommand(stop.New(streams))
+
+	o := newOptions(streams)
+	o.configFlags.AddFlags(cmd.Flags())
+
+	return cmd
+}