+++ pkg/nificlient/processor.go
@@ -24,3 +24,20 @@ func (n *nifiClient) UpdateProcessorRunStatus(
 
 	return &processor, nil
 }
+
+func (n *nifiClient) GetProcessor(id string) (*nigoapi.ProcessorEntity, error) {
+	// Get nigoapi client, favoring the one associated to the coordinator node.
+	client, context := n.privilegeCoordinatorClient()
+	if client == nil {
+		n.log.Error("Error during creating node client", zap.Error(ErrNoNodeClientsAvailable))
+		return nil, ErrNoNodeClientsAvailable
+	}
+
+	// Request on Nifi Rest API to get the processor informations
+	processor, rsp, body, err := client.ProcessorsApi.GetProcessor(context, id)
+	if err := errorUpdateOperation(rsp, body, err, n.log); err != nil {
+		return nil, err
+	}
+
+	return &processor, nil
+}