+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 - [PR #292](https://github.com/konpyutaika/nifikop/pull/292) - **[Operator/NifiCluster]** Modify RBAC kubebuilder annotations so NiFiKop works on OpenShift
 - [PR #292](https://github.com/konpyutaika/nifikop/pull/292) - **[Helm Chart]** Add Parameter for RunAsUser for OpenShift
 
+- [PR #291](https://github.com/konpyutaika/nifikop/pull/291) - **[Plugin]** Implementation on NiFiKop's plugin.
+- [PR #291](https://github.com/konpyutaika/nifikop/pull/291) - **[Operator/NifiConnection]** Implementation on NifiConnection controller.
+- 
 ### Changed
 
-- [PR #290](https://github.com/konpyutaika/nifikop/pull/290) - **[Operator/NifiCluster]** Change default sensitive algorithm
+- [PR #290](https://github.com/konpyutaika/nifikop/pull/290) - **[Operator/NifiCluster]** Change default sensitive algorithm.
 
 ### Fixed Bugs